Fix: Remove user info in userDetails on logout
Adds USER_DETAILS_RESET to remove user's information for the userDetails
state on logout. This fixes a bug where the previously logged in user's
name and email were shown when the next user to log in goes their
profile screen. I believe this bug would only happen if both users were
using the same device and the same browser session to log in and out.